+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## v1.4.4.dev
 
+#### Added:
+
+* Added support for shell completion. Completion is available for the commands
+  themselves and for several arguments, notably paths relative to the Dropbox folder and
+  choices from a set of fixed options. Use the command `maestral completion` to generate
+  shell completion scripts for bash, zsh or fish. For bash or zsh, save the returned
+  script in a location of you choice and source it in '~/.bashrc' or '~/.zshrc',
+  respectively. For fish, save the returned script at
+  '~/.config/fish/completions/maestral.fish'. Completion is available for the base
+  commands themselves and several arguments, notably paths relative to the Dropbox
+  folder and choices from a set of fixed options.
+
 #### Changed:
 
 * GUI dialogs to select a local Dropbox folder now ask for the actual folder name